+++ tests/unit/storage_test_utils.hpp
@@ -15,3 +15,8 @@
 #include "storage/v2/view.hpp"
 
 size_t CountVertices(memgraph::storage::Storage::Accessor &storage_accessor, memgraph::storage::View view);
+
+std::string_view StorageModeToString(memgraph::storage::StorageMode storage_mode);
+
+inline constexpr std::array storage_modes{memgraph::storage::StorageMode::IN_MEMORY_ANALYTICAL,
+                                          memgraph::storage::StorageMode::IN_MEMORY_TRANSACTIONAL};
